development/requirements.txt
SQLAlchemy>=2.0
alembic>=1.13
psycopg2-binary
